+++ tests/Get-OpenADAuthSupport.Tests.ps1
@@ -1,0 +1,11 @@
+. ([IO.Path]::Combine($PSScriptRoot, 'common.ps1'))
+
+Describe "Get-OpenADAuthSupport" {
+    It "It gets auth support details" {
+        $actual = Get-OpenADAuthSupport
+        $actual.Count | Should -BeGreaterThan 0
+        $actual | ForEach-Object {
+            $_ -is ([PSOpenAD.AuthenticationProvider]) | Should -BeTrue
+        }
+    }
+}