+++ tools/run-ci.sh
@@ -2,13 +2,19 @@
 
 IMAGE=${1}
 if [[ -z "${IMAGE}" ]]; then
-    echo "Must provide docker image to use for the test as first argument" 1>&2
+    echo "Must provide docker image to use for the test as the first argument" 1>&2
     exit 1
 fi
 
-BUILD_CONFIGURATION=${2}
+GSSAPI_PROVIDER=${2}
+if [[ -z "${GSSAPI_PROVIDER}" ]]; then
+    echo "Must provide GSSAPI provider [mit|heimdal] use for the test as the second argument" 1>&2
+    exit 1
+fi
+
+BUILD_CONFIGURATION=${3}
 if [[ -z "${BUILD_CONFIGURATION}" ]]; then
-    echo "Must provide build configuration (Debug|Release) as first argument" 1>&2
+    echo "Must provide build configuration (Debug|Release) as the third argument" 1>&2
     exit 1
 fi
 
@@ -31,17 +37,18 @@ fi
 function cleanup()
 {
     if [ "$( $DOCKER_BIN ps -q -f id=${DC_CONTAINER_ID} )" ]; then
-        $DOCKER_BIN kill "${DC_CONTAINER_ID}"
+        $DOCKER_BIN kill "${DC_CONTAINER_ID}" >/dev/null 2>&1
     fi
 
     $DOCKER_BIN network inspect "${NETWORK_NAME}" >/dev/null 2>&1 && \
-      $DOCKER_BIN network rm "${NETWORK_NAME}"
+      $DOCKER_BIN network rm "${NETWORK_NAME}" >/dev/null 2>&1
 }
 trap cleanup EXIT
 
 $DOCKER_BIN network inspect "${NETWORK_NAME}" >/dev/null 2>&1 || \
     $DOCKER_BIN network create --driver bridge "${NETWORK_NAME}"
 
+echo "Starting Samba DC container"
 DC_CONTAINER_ID=$( $DOCKER_BIN run \
     --detach \
     --rm \
@@ -54,6 +61,20 @@ DC_CONTAINER_ID=$( $DOCKER_BIN run \
     --network-alias "dc.${REALM,,}" \
     debian:11 /bin/bash /tmp/PSOpenAD/tools/setup-samba.sh )
 
+echo "Getting Samba DC container IP"
+DC_IP=$( $DOCKER_BIN inspect -f \
+    '{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}' \
+    "${DC_CONTAINER_ID}"
+)
+
+echo "Waiting for Samba to come online"
+$DOCKER_BIN exec \
+    --interactive \
+    --tty \
+    "${DC_CONTAINER_ID}" \
+    /bin/bash -c "until pidof samba >/dev/null 2>&1; do sleep 1; done"
+
+echo "Starting test container"
 $DOCKER_BIN run \
   --rm \
   --interactive \
@@ -61,14 +82,18 @@ $DOCKER_BIN run \
   --workdir /tmp/PSOpenAD \
   --env AD_REALM="${REALM^^}" \
   --env AD_PASSWORD="${PASSWORD}" \
+  --env GSSAPI_PROVIDER="${GSSAPI_PROVIDER}" \
   --env BUILD_CONFIGURATION="${BUILD_CONFIGURATION}" \
   --hostname "app.${REALM,,}" \
   --network "${NETWORK_NAME}" \
   --network-alias app \
   --network-alias "app.${REALM,,}" \
+  --dns "${DC_IP}" \
   "${IMAGE}" /bin/bash -e -c 'source /dev/stdin' << 'EOF'
 
 source ./tools/lib.sh
+
+rm -f output/PSOpenAD*.nupkg
 
 lib::setup::system_requirements
 lib::setup::gssapi